breaking out convenience validators
