+++ Sources/Vapor/Validation/Convenience/Email.swift
@@ -1,12 +1,11 @@
-// Might have to tweak for linux
 import Foundation
 
 public class Email: ValidationSuite {
     public static func validate(input value: String) throws {
-        guard
-            // Thanks Ben Wu :)
-            value.range(of: ".@.+\\..", options: .regularExpressionSearch) != nil
-            else {
+        // Thanks Ben Wu :)
+        let range = value.range(of: ".@.+\\..",
+                                options: .regularExpressionSearch)
+        guard let _ = range else {
             throw error(with: value)
         }
     }
